Update ev-node version and refactor cometcompat to adapter
- Bump github.com/evstack/ev-node to
v1.0.0-beta.2.0.20250821181753-974aa15383de - Move cometcompat helpers
to adapter package and update imports - Remove cometcompat package and
related files - Add providers.go to adapter for signature and validator
hash providers - Refactor usage across codebase to use adapter package
functions
