+++ .github/workflows/integration_test.yml
@@ -16,7 +16,7 @@ jobs:
     runs-on: ubuntu-latest
     timeout-minutes: 30
     env:
-      EVNODE_VERSION: "v1.0.0-beta.2.0.20250819153439-29b8ddf5a753"
+      EVNODE_VERSION: "v1.0.0-beta.2.0.20250821181753-974aa15383de"
       IGNITE_VERSION: "v29.3.1"
       IGNITE_EVOLVE_APP_VERSION: "main" # use tagged when apps has tagged (blocked on things)
       EVOLVE_IMAGE_REPO: "evolve-gm"
@@ -59,7 +59,7 @@ jobs:
     timeout-minutes: 30
     env:
       DO_NOT_TRACK: true
-      EVNODE_VERSION: "v1.0.0-beta.2.0.20250819153439-29b8ddf5a753"
+      EVNODE_VERSION: "v1.0.0-beta.2.0.20250821181753-974aa15383de"
       IGNITE_VERSION: "v29.3.1"
       IGNITE_EVOLVE_APP_VERSION: "main" # use tagged when apps has tagged (blocked on things)
     outputs:
@@ -266,7 +266,7 @@ jobs:
       GMD_HOME: ${{ needs.liveness.outputs.gmd_home }}
       HERMES_VERSION: "v1.13.1"
       GAIA_VERSION: "v25.1.0"
-      EVNODE_VERSION: "v1.0.0-beta.2.0.20250819153439-29b8ddf5a753"
+      EVNODE_VERSION: "v1.0.0-beta.2.0.20250821181753-974aa15383de"
       EVNODE_DA_VERSION: "v1.0.0-beta.1"
     steps:
       - name: Set up Go
