+++ .github/workflows/migration_test.yml
@@ -13,7 +13,7 @@ jobs:
     timeout-minutes: 45
     env:
       DO_NOT_TRACK: true
-      EVNODE_VERSION: "v1.0.0-beta.2.0.20250819153439-29b8ddf5a753"
+      EVNODE_VERSION: "v1.0.0-beta.2.0.20250821181753-974aa15383de"
       IGNITE_VERSION: "v29.3.1"
       IGNITE_EVOLVE_APP_VERSION: "main" # use tagged when apps has tagged (blocked on things)
     steps:
